services/python/pyproject.toml
[project]
name = "zeniipo-python-sidecar"
version = "1.0.0"
description = "FastAPI sidecar for Zeniipo: Monte Carlo, Council validation, batch Claude agents"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Zeni Holdings", email = "ops@zenidigital.com" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.27.0",
  "anthropic>=0.39.0",
  "supabase>=2.9.0",
  "numpy>=2.1.0",
  "scipy>=1.14.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
